+++ app/src/main/java/org/mozilla/focus/telemetry/TelemetryWrapper.kt
@@ -3956,7 +3956,7 @@ object TelemetryWrapper {
     @JvmStatic
     fun startDownloadFile(
         downloadId: String,
-        fileSize: Double,
+        fileSize: Long,
         isValidSSL: Boolean,
         isSupportRange: Boolean
     ) {
@@ -3987,7 +3987,7 @@ object TelemetryWrapper {
     @JvmStatic
     fun endDownloadFile(
         downloadId: Long,
-        fileSize: Double,
+        fileSize: Long,
         progress: Double,
         status: Int,
         reason: Int
